Add server-side check for existing preset

+++ server/presetModel.js
@@ -3,12 +3,16 @@ const path = require('path');
 
 module.exports = {
   create(synthData, callback) {
-    fs.writeFile(path.resolve(__dirname, `./presets/${synthData.name || 'default'}.websynth.json`), JSON.stringify(synthData.synthesizer), (err) => {
-      if (err) {
-        callback(err);
-      } else {
-        callback(null);
-      }
-    });
+    if (synthData.overwrite === false && fs.existsSync(path.resolve(__dirname, `./presets/${synthData.name || 'default'}.websynth.json`))) {
+      callback('exists');
+    } else {
+      fs.writeFile(path.resolve(__dirname, `./presets/${synthData.name || 'default'}.websynth.json`), JSON.stringify(synthData.synthesizer), (err) => {
+        if (err) {
+          callback(err);
+        } else {
+          callback(null);
+        }
+      });
+    }
   }
 };
